requirements.txt
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
peewee==3.9.6
pycparser==2.19
pysqlite3==0.2.1
python-dotenv==0.10.3
pytz==2019.1
redis==3.2.1
six==1.12.0
Werkzeug==0.15.4
WTForms==2.2.1
